refactor: extract ProductCard and ProductCardSkeleton into separate files for reusability
