+++ src/components/ProductCard/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductCard } from './ProductCard';
