+++ src/components/ProductCardSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductCardSkeleton } from './ProductCardSkeleton';
