+++ src/pages/Home/Home.module.css
@@ -15,92 +15,6 @@
   padding: 1rem 0;
 }
 
-.card,
-.skeleton {
-  display: flex;
-  flex-direction: column;
-  background-color: var(--color-surface);
-  border-radius: 12px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
-}
-
-.card {
-  color: var(--color-text-primary);
-  cursor: pointer;
-  transition: all 250ms ease-in-out;
-
-  &:hover {
-    box-shadow: 0 4px 16px rgba(0, 0, 0, 0.15);
-    transform: scale(1.02);
-  }
-}
-
-.imageContainer {
-  position: relative;
-  width: 100%;
-  padding-top: 75%;
-  background-color: var(--color-background);
-  border-bottom: 1px solid var(--color-border);
-  overflow: hidden;
-  border-top-left-radius: 12px;
-  border-top-right-radius: 12px;
-}
-
-.image {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  top: 0;
-  left: 0;
-  object-fit: cover;
-}
-
-.content {
-  padding: 1rem 1.5rem;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-
-.category {
-  color: var(--color-text-secondary);
-  font-weight: 400;
-  margin-bottom: 0.25rem;
-}
-
-.name {
-  font-weight: 600;
-  font-size: 1.3rem;
-  margin-bottom: 1rem;
-}
-
-.footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: auto;
-}
-
-.price {
-  font-weight: 700;
-  font-size: 1.5rem;
-}
-
-.cartButton {
-  color: white;
-  background-color: var(--color-primary);
-  font-size: 1rem;
-  font-weight: 600;
-  padding: 0.5rem 1rem;
-  border-radius: 8px;
-  transition: all 250ms ease-in-out;
-
-  &:hover {
-    background-color: #1b6ebc;
-    transform: scale(1.05);
-  }
-}
-
 .noMatches {
   height: 60vh;
   display: flex;
